DAGs/google cloud/ingest/requirements.txt
requests==2.30.0
google-cloud-storage==2.10.0
python-dotenv==1.0.0